ODP API Testing
Modified odp/test/Makefile to compile api testing source files which
are newly added to odp/test/api_test. Since all the CFLAGS, LDFLAGS
and libraries are common to both sample apps and API testing, modified
the current Makefile and didn't add a new one.

TODO: As and when the necessity comes, we should have three makefiles.
First for Libraries and flags, second for sample apps and third for API
testing. Both second and third should include first Makefile.

Currently kept the atomic and shared memory binaries separate to
provide the basic infrastructure. As and when new set of API's are added,
there will be more binaries.

Signed-off-by: venkatesh vivekanandan <venkatesh.vivekanandan@linaro.org>