tree-wide: introduce SPDX Unique License Identifiers
ODP includes big blocks of License headers in all files.  This not
only blows up the source code with mostly redundant information, but
also makes it very difficult to generate License Clearing Reports.
An additional problem is that even the same licenses are referred to
by a number of slightly varying text blocks (full, abbreviated, different
indentation, line wrapping and/or white space, with obsolete address
information, ...) which makes automatic processing a nightmare.

To make this easier, such license headers in the source files will be
replaced with a single line reference to Unique License Identifiers
as defined by the Linux Foundation's SPDX project [1].  I.e,
in a source file the full BSD 3-clause header text will be
replaced by a single line:

       SPDX-License-Identifier:        BSD-3-Clause

We use the SPDX Unique License Identifiers here; these are available
at [2].

Note: From the legal point of view, this patch is supposed to be only
a change to the textual representation of the license information,
but in no way any change to the actual license terms. With this patch
applied, all files will still be licensed under the same terms they
were before.

[1] http://spdx.org/
[2] http://spdx.org/licenses/

Signed-off-by: Kim Phillips <kim.phillips@linaro.org>