Accept short frames from netmap software rings
Since packets from the network stack are mostly valid, we decided to pad them to
a minimum of OD_ETH_LEN_MIN. The odp_packet_parse function is intended to be
used for parsing frames for physical wired ethernet devices which are supposed
to conform to IEEE 802.3 minimum frame length. Padding frames from the network
stack (SW ring) is the prefered choice at the moment.

Signed-off-by: Ciprian Barbu <ciprian.barbu@linaro.org>