Fix drop_err_pkts to correctly discard bad packets
The drop_err_pkts function is intended to discard packets that input
parsing has marked as containing errors (e.g. short frame). It correctly
identifies and frees bad packets but an error in the way pkt_tbl is
updated means bad packet handles aren't actually removed, later use of
the stale packet handle results in a SEGV. The pkt_cnt is decremented
though, so for each bad packet a good packet (within the same burst) is
dropped and leaked.

Signed-off-by: Stuart Haslam <stuart.haslam@arm.com>