Fix mismatched size of ODP_PACKET_OFFSET_INVALID
ODP_PACKET_OFFSET_INVALID is defined as ((size_t)-1) but the existing
implementations both store offsets using a uint32_t, which breaks things
in a few places on 64-bit systems. There's some explicit casting in the
implementations but it's not consistent and the odp_packet_l2/3/4
functions can return a junk ptr if packet parsing didn't get far enough
to set the offset correctly.

Signed-off-by: Stuart Haslam <stuart.haslam@arm.com>