odp_packet_socket: define PACKET_FANOUT if needed
PACKET_FANOUT and friends is only defined post kernel 3.1,
meaning a build against older kernel header fails.

By providing a backwards compatible define for older kernels
we allow the build to work, if run time attempts of using
the missing functionality on older kernels to degrade to
a run-time EINVAL error in init time if the specific feature
is used.

Signed-off-by: Gilad Ben-Yossef <giladb@ezchip.com>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>