odp_packet_socket: add backwards compatible sendmmsg wrapper
RHEL / CentOS 6.2 and above have sendmmsg kernel support but
no libc support.

Also, it is quite easy to emulate sendmmsg on top of sendmsg.

The following patch provides a sendmmsg wrapper as a weak symbol
that provides both support to call the syscall even if libc support
is missing, as well as an emulation in case no kernel support is
there.

This lets us drop the dependency on kernel version 3.0.

Signed-off-by: Gilad Ben-Yossef <giladb@ezchip.com>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>