validation: crypto: remove data offset workaround
Current crypto tests calculate range offsets expecting a bug in crypto
implementation: data offsets are applied relatively to a start of a
buffer instead of start of data in a packet. This bug is fixed in
linux-generic by recent buffer/packet rework patches, so all crypto
tests fails now. Remove this workaround.

Signed-off-by: Taras Kondratiuk <taras.kondratiuk@linaro.org>
Reviewed-by: Alexandru Badicioiu <alexandru.badicioiu@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>