validation: pktio: initial pktio tests
This is primarily sunny day basic functional testing, all tests are
single threaded.

On platforms other than linux-generic testing will be performed using
the "loop" interface by default.

For linux-generic, since we don't have a working "loop" device, a
wrapper script is run that attempts to create a pair of virtual
Ethernet interfaces for testing. If creating the interfaces fails
the test is skipped and "make check" reports it as such.

Signed-off-by: Stuart Haslam <stuart.haslam@arm.com>
Reviewed-by: Bill Fischofer <bill.fischofer@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>