api: odp_packet_io.h: make pktio handles single instance per device
This patch changes odp_pktio_open() to ensure a one-to-one relationship
between a pktio device an a pktio handle. Attempts to open a device
which is already open will fail, setting errno to -EEXIST to indicate
the reason for the failure. A new API odp_pktio_lookup() is added which
can be used to obtain a handle to an already open device.

Signed-off-by: Stuart Haslam <stuart.haslam@arm.com>
Reviewed-by: Bill Fischofer <bill.fischofer@linaro.org>
Signed-off-by: Mike Holmes <mike.holmes@linaro.org>