linux-generic: pktio: add loopback device support
Create a new linux-generic pktio type ODP_PKTIO_TYPE_LOOPBACK which
will be used when the "loop" device is opened. A queue is used to
connect the pktio output queue to the input queue. Packets sent to the
loopback device are parsed and classified before making it to the input
queue. A fixed MAC address is used and the other control functions (MTU
and promisc mode) are essentially just dummies.

As this doesn't need root access, the odp_pktio_run script is modified
to use the "loop" device when not root. The veth method is still used
when running as root (on linux-generic) so that the socket code
actually gets tested.

Signed-off-by: Stuart Haslam <stuart.haslam@arm.com>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>