example: odp_timer_test.c: odp_timer_test aborts
Since the scheduler does not guarantee fairness in scheduling, threads must
time out and check if the example is still running. If so, just restart the
scheduling, not the whole example loop which has (undesired) side effects.
The bug was caused by restarting the example loop which then reset the last
used timer with an old tick value, this could eventually fail (because that
tick value was in the past) with the too late error message.

https://bugs.linaro.org/show_bug.cgi?id=1068

Reported-by: anders.roxell@linaro.org
Reported-by: ola.dahl@enea.com
Signed-off-by: Ola Liljedahl <ola.liljedahl@linaro.org>
Reviewed-and-tested-by: Bill Fischofer <bill.fischofer@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>