linux-generic: pool set seg_len to 0
gcc 4.9 compiles current code fine, but gcc 4.7 fails
with:
odp_pool.c:293:20: error: 'seg_len' may be used
uninitialized in this function
Problem found with gcc 4.7 and make distcheck command which added
additional flags for odp compilation. If newer gcc does not warn that
variable can be used uninitialized (not for packet case in this function),
it looks like compiler bug. Init it to 0 because pool creation is not
performance function.

Reviewed-by: Bill Fischofer <bill.fischofer@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>