api: Make odp_packet_len inline
This function is in some very hot paths, often called for every packet. Simply
the overhead of the function call could be quite big, inlineing can provide up
to 10% performance gain. Unfortunately we can't show in the API the internals,
so at compile time we calculate the offset for the packet length.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>