performance: pktio: add throughput test
A self-contained throughput test using the loop interface.

A number of transmit and receive workers are run on separate cores,
the transmitters generate packets at a defined rate and the receivers
consume them. Each test iteration runs for a fixed period, at the
end of the iteration it is verified that the number of packets
transmitted was as expected and that all transmitted packets were
received.

It's also possible to specify an alternative loopback device or a
pair of connected devices via the command line.

Signed-off-by: Stuart Haslam <stuart.haslam@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>