pool: calculate cache size before pool creation
According to the DPDK API, this value must be equal or smaller than
RTE_MEMPOOL_CACHE_MAX_SIZE. Also, it is recommended to be a divisor of the
number of elements.
This algorithm tries to find the greatest divisor which is smaller than the
max allowed, if there isn't any, the max is used.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>