pktio: remove send_pkt_dpdk()
Apart from calling rte_eth_tx_burst the only thing it does is to copy the
pkt_table[] array into an another one. As far as I can determine DPDK doesn't
touch this array of pointers, so it's not necessary to copy it. I've merged the
rest of it into odp_pktio_send()

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>