configure.ac: include configure.m4 if platform explicitly selected
cherry-pick linux-generic commit: 920c3db
configure script selects linux-generic as default platform and includes
platfrom specific configure scripts: ./platform/linux-generic/m4/configure.m4
In case if --with-platfrom=linux-generic was provided to configure platform
specific file is not included and that branch of code does not run.
This change has to be done for each platform since we have there pktio_env
scripts but m4_include macro does not recognize bash variables. So that there
is if check.

Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>
Reviewed-by: Anders Roxell <anders.roxell@linaro.org>
Reviewed-by: Ciprian Barbu <ciprian.barbu@linaro.org>
Reviewed-and-tested-by: Mike Holmes <Mike.holmes@linaro.org>