pool: Refactor cache_size algorithm
This patch tries to find the greatest divisor of n, which is:

- not itself
- smaller or equal than RTE_MEMPOOL_CACHE_MAX_SIZE
- cache_size * 1.5 smaller or equal than n

The third point implies the first point, and the fact that it should be a
divisor, implies the third point as well. That's why j starts at least from 2.
There is an extra check after the calculation to make sure the value conforms
the requirements.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>
Reviewed-by: Bill Fischofer <bill.fischofer@linaro.org>