packet_io: change TX completion
Originally we flushed the completed buffers from the current interface before
calling receive, but buffers are often held in other queues.
Instead we check if receive was unsuccessful and the attached pool doesn't
have any buffers for receiving. In that case we go through the list of all
pktio interfaces (first the current), and try to flush out completed
descriptors until we get something.
rte_mempool_available() is based on rte_mempool_count(), but it only count
the main ring and the current thread's cache, because the others are not
interesting for us.
This method doesn't scale very well when there are a lot of interfaces, and/or
they change very often. We need to refactor the underlying pktio infrastructure
first to enable fast iteration through the list of existing interfaces.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>