linux-generic: timer: set timer queue to ODP_QUEUE_INVALID on init
The set_next_free() checks if timer queue is not set, that is
ODP_QUEUE_INVALID. In case of linux-generic the invalid queue handle
is 0. By coincidence, after allocation, the timer queue handle is
also 0, by this way the ODP_ASSERT is masked.
In case ODP_QUEUE_INVALID is another from 0 the test generates error.
It's not correct, the code shouldn't have such kind dependency from
macro definition, and it prevents to re-use this code for another
platforms with different ODP_QUEUE_INVALID. So, fix it.

Reviewed-by: Ola Liljedahl <ola.liljedahl@linaro.org>
Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>