validation: cpumask: def_worker and def_control fixes
The odp_cpumask_def_worker() and odp_cpumask_def_control() APIs are used
to populate a mask with the CPUs available for running threads of the
same type. This does not necessarily mean those CPUs are dedicated to
running threads of that type, on a particular platform the same CPU may
appear in both the control and worker set.

The test now verifies that the APIs;

 - populate the mask with at least 1 CPU
 - return a value that matches the number of CPUs set in the mask
 - don't return a value greater than the total available CPUsa
 - don't return a value greater than the requested number of CPUs

Change-Id: I31d6875df458ef51aceb37949270179d612457b8
Signed-off-by: Stuart Haslam <stuart.haslam@linaro.org>
Reviewed-by: Christophe Milard <christophe.milard@linaro.org>