example: timer: delete races while termination
Current implementation has at least two races that lead to several issues:

- gbls->remain can overflow. One thread can decrement remain counter to 0.
While another can decrement it once again and it will be > 0. After
what some thread will loop very long time ...

- Several threads can terminate the same timer and as result the same
event. After out from the main loop a thread terminates a last timer it
used. But a last timer saved in ttp for a thread can be received in
another thread. So after leaving the main loop two threads can hold the
same timer.

- Some timer cannot be freed as several threads try to delete the same
timer, as result one of the timer/tmo stay not freed after termination.

- The test can send more events that requested. The receiving of requested
number of tmos doesn't mean the test sent the same number. It rather sent
more.

This patch is intended to fix above drawbacks.

The termination path must follow the next things:

- An event can be in the following places: in a timer (waiting to be
scheduled), in a queue for a thread to be scheduled, received in the
main loop.

- An event "holds" a timer, so when we receive an event we can delete it's
timer.

- a thread cannot delete timer w/o an event as it doesn't know who is
owner of the event (and obvious the timer).

- a thread shouldn't send events more than requested.

- all threads have to be "held" in the loop till a last received event.
The scheduler can assign event for any of the threads, so one thread can
receive two last events for example.

According to above, added several improvements:
- don't send more timeouts that supposed to receive
- free timer and tmo for a last received tmos = num of threads.
- leave the main loop only if a last tmo/timer is free.

Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>