example: timer: use ability to indicate that event was lost
In case if for some reason one of the TMOs was lost, the test will
be polling forever in order to take it and delete. It can be used to
indicate that event was lost. Just abort the test if it cannot receive
one of the "last" TMOs for some period.

Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>