linux-dpdk: pktio: use packet input field for odp_packet_input
Fixes https://bugs.linaro.org/show_bug.cgi?id=1700
The way odp_packet_input was implemented made use of a per-pool pktio handle
which meant that a pool could only be used by a single pktio. This was not
correct from an API perspective and was probably old code remaining from the
early days of odp-dpdk. As an argument for this there was already existing code
that set a per-packet pktio handle, probably merged from upstream odp, but not
actually used by odp_packet_input.

Signed-off-by: Ciprian Barbu <ciprian.barbu@enea.com>
Reviewed-by: Bill Fischofer <bill.fischofer@linaro.org>
Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>