linux-dpdk: fix odp_cpumask_def_worker
According to the API documentation odp_cpumask_def_worker sets up to 'num' CPUs
and returns the count actually set or all available CPUs when num is 0. The
former implementation was always returning the number of slave lcores, which is
typically cpu_count -1. Note that if num == 0 the loop will iterate through all
available slave lcores, thus conforming to the definition.

Signed-off-by: Ciprian Barbu <ciprian.barbu@enea.com>
Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>