linux-dpdk: test: pktio: fix pktio_run PATH setup
When the build of tests outputs into an another directory (e.g. ./testdir)
and 'make check' is running from there, pktio_run will run from the original
source dir, so $(dirname $0)/../../../../test/validation/pktio will point to
the original source as well. But pktio_main is in the build output dir.
This patch adds that to the path. This should particularly happen in CI.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>