api: pool: add packet user area initializer for pool creation parameters
Applications can preset certain parts of the packet user area, so when that
memory will be allocated it starts from a known state. If the platform
allocates the memory during pool creation, it's enough to run the
constructor after that. If it's allocating memory on demand, it should
call the constructor each time.
Porting applications to ODP can benefit from this. If the application can't
afford to change its whole packet handling to ODP, it's likely it needs to
maintain its own metadata in the user area. And probably it needs to set
constant fields in that metadata e.g. to mark that this is an ODP packet,
and/or store the handle of the packet itself.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>
Reviewed-by: Balasubramanian Manoharan <bala.manoharan@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>