linux-dpdk: pktio: handle vector PMDs
The ixgbe PMD introduced vector operations on the RX path, and the plan is
to change the other drivers into this direction as well. Unfortunately
that means these new drivers mandate a minimum burst size for receive,
otherwise they don't return any packets. As ODP can't impose such a
requirement on applications, we have to deal with it during the receive
function.
Currently ixgbe requires 32 buffers, it is expected to change to 4 later
on. The rest of the packets in the burst are dropped, and a debug message
is printed.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>