checkpatch.conf allow split strings
ODP has proven a need to accommodate strings that cannot be accommodated on
an 80 character line. Allow spilt strings as the preferred way of
solving the issue so that check patch rules do not make code
submission impossible. Do this in a way that does not require merging
local changes to check patch when updating to a new upstream version.

https://bugs.linaro.org/show_bug.cgi?id=1746

Signed-off-by: Mike Holmes <mike.holmes@linaro.org>
Reviewed-by: Bill Fischofer <bill.fischofer@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>