api: schedule: revise definition of ordered locks
Revise the definition of odp_schedule_order_lock() and
odp_schedule_order_unlock(). These APIs now take as an argument the
index value of the ordered lock within the current context to be locked
or unlocked. Because the API is changed, this patch also updates the
linux-generic implementation of these APIs as well as the scheduler
validation test that uses them.

Signed-off-by: Bill Fischofer <bill.fischofer@linaro.org>
Reviewed-by: Petri Savolainen <petri.savolainen@nokia.com>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>