linux-generic: sysinfo: make the model_str per-CPU data
For AMP system such as ARM big.LITTLE, cores are heterogeneous, and the
model_str for each core may be different too, so this patch changes the
model_str to data array model_str[] to contain data for each different
core, while for the common SMP system, we can simply use the model_str[0]
to contain data for all cores because they are all same, but if like, we
can fill each item in the data array too.

The new API to get each model_str for AMP system will be added later.

Signed-off-by: Hongbo Zhang <hongbo.zhang@linaro.org>
Reviewed-by: Petri Savolainen <petri.savolainen@nokia.com>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>