validation: add ability to mark tests inactive
Add the ability for individual test cases to be marked as inactive,
either unconditionally or based on the result of a check function which
is called at test init time. Inactive tests are registered with CUnit
and are reported as inactive in test reports without being executed.

All uses of CU_TestInfo have been replaced with odp_testinfo_t in order
to carry the conditional function pointer in addition to the CU_TestInfo
struct, and some new macros added to make things a bit easier;

_CU_TEST_INFO() is renamed to ODP_TEST_INFO() but remains functionally
the same, these tests will be run unconditionally.

ODP_TEST_INFO_INACTIVE() can be used to define an inactive test case.
There's no good reason to do this in any of the modules under
test/validation/ but the intention is to allow this to be done on a
per-platform basis from within the platform/<platform>/test/ tree.

ODP_TEST_INFO_CONDITIONAL() takes an additional parameter which is a
pointer to a function to call at test suite initialisation time to
determine whether that test should be marked as inactive.

Signed-off-by: Stuart Haslam <stuart.haslam@linaro.org>
Reviewed-by: Christophe Milard <christophe.milard@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>