validation: add odp_cunit_update() to modify registered tests
Add an odp_cunit_update() function to allow some attributes of previously
registered tests to be modified, such as whether it's active or not.

Previously, registering and running tests was done in a single step;

odp_cunit_run(testsuites);

But this is now split;

odp_cunit_register(testsuites);
odp_cunit_update(testsuites_updates);
odp_cunit_run();

The odp_cunit_update() is optional and none of the currently defined
tests use it, so there's no functional change in this patch.

Signed-off-by: Stuart Haslam <stuart.haslam@linaro.org>
Reviewed-by: Christophe Milard <christophe.milard@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>