linux-generic: pktio: add pcap pktio type
Create a new pktio type that allows for reading from and writing to a
pcap capture file. This is intended to be used as a simple way of
injecting test packets into an application for functional testing and
can be used as it is with some of the existing example applications.

To use this interface the name passed to odp_pktio_open() must begin
with "pcap:" and be in the format;

 pcap:in=test.pcap:out=test_out.pcap:loops=10

   in      the name of the input pcap file. If no input file is given
           attempts to receive from the pktio will just return no
           packets.
   out     the name of the output pcap file. If no output file is
           given any packets transmitted over the interface will just
           be freed.
   loops   the number of times to iterate through the input file, set
           to 0 to loop indefinitely. The default value is 1.

Signed-off-by: Stuart Haslam <stuart.haslam@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>