example:generator : Fix data race condition
The counters.seq counter is used to check if the configured number of
packets was processed. There is a race condition between the counter
incrementation time and its value testing time. If code is running on
multiple CPUs it is possible the application send more packets than
expected (with number of CPUs - 1). A separate counter must be used
for the processed packets.

Signed-off-by: Grigore Ion <ion.grigore@freescale.com>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>