validation:classification: Add fix for classification tests
odph_ipv4_csum_update should be used to update the checksum inside a pkt,
as it is used in all the other examples and tests different from
classification. Thus the prototype of the function should return void,
because the intention was to update a value not to return something. These
being said it is wrong(on LE platforms) to do a cpu_to_be conversion
and an assignment operation in classification_test(this is already done
inside).

Signed-off-by: Grigore Ion <ion.grigore@freescale.com>
Reviewed-by: Balasubramanian Manoharan <bala.manoharan@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>