validation: pktio: pass interface index rather than name
Avoid the need to pass the interface name to the create_pktio() function
as it can be derived from the index which must also be passed. This
means the caller can't make the mistake of passing a mismatched name and
index - there was one instance of this which is now fixed.

Signed-off-by: Stuart Haslam <stuart.haslam@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>