linux-generic: queue: use locking hierarchy for ordered queueing
Enqueueing to ordered queues requires two locks (source and target
queues). Since any queue can be either source or target, queues do not
have a natural locking hierarchy. Create one by using the address of
the qentry as the lock order.

This addresses the aspect of Bug
https://bugs.linaro.org/show_bug.cgi?id=1879
relating to deadlock in unicore systems.

Suggested-by: Nicolas Morey-Chaisemartin <nmorey@kalray.eu>
Signed-off-by: Bill Fischofer <bill.fischofer@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>