linux-dpdk: errno: use rte_errno instead of __odp_errno
It is DPDK's errno variable. odp_pktio_send() also makes sure that it is set
even if the PMD doesn't do that.
The macro in odp_buffer_inlines.h is a workaround to make sure linux-generic
timer implementation doesn't have a linking issue. It would be slightly
better to have it in a local copy of odp_internal.h, but having it in sync
with linux-generic is more important.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>