validation: remove strict dependency on CUnit 2.1-3
This makes it a bit easier to get a new system set up for running the
unit tests. There has always been a dependency on using a version of
CUnit >= 2.1-3, using an older version would result in an error:

error: initialization discards ‘const’ qualifier from pointer target type [-Werror]

caused by CUnit's CU_TestInfo structure taking a char * for the test
name rather than a const char *, and we always pass string literals as
names. Version 2.1-3 of CUnit changed this but that's still not packaged
with most distributions so we end up needing to build it from source
most of the time. We now don't use the CU_TestInfo structure directly in
any of our tests so a couple of small changes in the common code make
the issue go away.

Tested with CUnit versions 2.1-2.dfsg-1 (packaged with Ubuntu 14.04) and
2.1-3 (which most people will still be using).

Signed-off-by: Stuart Haslam <stuart.haslam@linaro.org>
Reviewed-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>