linux-generic: cpumask: warn that CPU0 is used by control and worker thread
By default, all control threads execute on CPU 0 and
odp_cpumask_default_control() returns this CPU. For reasons of
performance, control and worker CPU's shouldn't overlap but for some
scenarios it can be desirable to utilize all CPU's for worker threads.
Thus we leave the decision of CPU allocation to the user but report
when a CPU is used for both control and worker threads.

Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>
Reviewed-by: Ola Liljedahl <ola.liljedahl@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>