validation: ability to specify test install directory
Previously if a build had been configured with validation tests enabled
(--enable-test-vald) a subsequent "make install" would install the test
binaries and scripts into /usr/local/bin without any way of separating
them from the rest of the installed artefacts. This patch changes the
default behaviour to not install test binaries or scripts and adds a
configure parameter (--with-testdir) to explicitly enable installation.

These changes are required in order to be able to package the tests
separately (e.g. in OpenEmbedded) and also to enable post-install
testing.

Signed-off-by: Stuart Haslam <stuart.haslam@linaro.org>
Reviewed-by: Anders Roxell <anders.roxell@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>