linux-generic: validation: add run-test script for post install testing
Add a run-test script that replicates some of the "make check"
functionality without depending on a whole load of Makefile gubbins.
This makes it easier to perform on-target testing following a
cross-build (as we want to do in CI) and gives a simple way to verify
that all the required components do actually get installed correctly,
using "make installcheck".

./configure --enable-test-vald --with-testdir
DESTDIR=$(pwd)/test-install make install installcheck

Signed-off-by: Stuart Haslam <stuart.haslam@linaro.org>
Reviewed-by: Anders Roxell <anders.roxell@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>