queue: fix memory corruption in reorder_enq()
reorder_prev is set to the address of the pointer origin_qe->s.reorder_head,
which is wrong. If the linked list was empty, that won't be corrected, and
reorder_prev->next points to the adjacent queue entry's status field. If that
entry is used, that queue's metadata will be corrupted.
This was found by running the chaos scheduler test with ODP-DPDK.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>
Reviewed-by: Bill Fischofer <bill.fischofer@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>