linux-dpdk: init: pass only one core through -c parameter
The others would be used for DPDK lcore threads, which are not accessible
for an ODP application. It requires ODP thread init to set up lcore_role,
though it's not widely used in other DPDK libs.
Also reset the original affinity of the calling thread, as DPDK limits it
down to one core.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>