linux-dpdk: cpumask: go back to linux-generic cpumask implementation
The only reason for a separate one is because odp_cpumask_default_worker()
goes through the lcore list to get this data for us. But with the master
thread's affinity fixed there is no need for that, the linux-generic way of
using the thread affinity is good for linux-dpdk as well.

Signed-off-by: Zoltan Kiss <zoltan.kiss@linaro.org>